Layout updates based on wireframe changes discussion
Have dashboard and a Analytics tabs
